Move GetDdevEnvironment() to separate package
